correct missed improper int size
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@41563 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/drivers/graphics/radeon_hd/radeon_hd_private.h
@@ -32,7 +32,7 @@ struct radeon_info {
 	area_id			shared_area;
 
 	const char*		device_identifier;
-	uint32			device_chipset;
+	uint16			device_chipset;
 };
